+++ design documents/readme.txt
@@ -1,0 +1,2 @@
+WARNING, SPOLIERS PRESENT
+IF YOU NEED HINTS, HERE WORKS, but you have been warned.
